fix(crux-ui): split the slowest deployment and image-config e2e test files
