+++ web/crux-ui/e2e/with-login/deployment/deployment.spec.ts
@@ -1,7 +1,7 @@
 import { ProjectType } from '@app/models'
 import { expect, Page, test } from '@playwright/test'
-import { DAGENT_NODE, NGINX_TEST_IMAGE_WITH_TAG } from '../utils/common'
-import { createNode } from '../utils/nodes'
+import { DAGENT_NODE, NGINX_TEST_IMAGE_WITH_TAG } from '../../utils/common'
+import { createNode } from '../../utils/nodes'
 import {
   addDeploymentToVersion,
   addDeploymentToVersionlessProject,
@@ -9,7 +9,7 @@ import {
   addImageToVersionlessProject,
   createProject,
   createVersion,
-} from '../utils/projects'
+} from '../../utils/projects'
 
 const setup = async (
   page: Page,
